pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.1",
    "keyring>=25.6.0",
    "keyrings-alt>=5.0.2",
    "pillow>=11.1.0",
    "pyperclip>=1.9.0",
    "pystray>=0.19.5",
]
